chore: make EthProvider seperate class, add babel plugins for Torus module
