+++ config-overrides.js
@@ -4,7 +4,10 @@ const path = require('path')
 module.exports = override(
     addPostcssPlugins([require('tailwindcss'), require('postcss-preset-env')({ stage: 1 })]),
     ...addBabelPlugins(
-        'babel-plugin-react-native-web'
+        'babel-plugin-react-native-web',
+        '@babel/plugin-proposal-optional-chaining',
+        '@babel/plugin-proposal-nullish-coalescing-operator',
+        '@babel/plugin-syntax-bigint'
         // '@babel/plugin-proposal-class-properties',
     ),
     ...addBabelPresets('@babel/preset-flow', '@babel/preset-react', '@babel/preset-typescript'),
