+++ package.json
@@ -27,6 +27,9 @@
         "@apollo/react-components": "^3.1.5",
         "@apollo/react-hooks": "^3.1.5",
         "@babel/core": "^7.15.8",
+        "@babel/plugin-proposal-nullish-coalescing-operator": "^7.18.6",
+        "@babel/plugin-proposal-optional-chaining": "^7.20.7",
+        "@babel/plugin-syntax-bigint": "^7.8.3",
         "@babel/preset-flow": "^7.18.6",
         "@babel/preset-react": "^7.18.6",
         "@babel/preset-typescript": "^7.18.6",
@@ -213,6 +216,7 @@
         "@web3-onboard/coinbase": "^2.0.10",
         "@web3-onboard/injected-wallets": "^2.0.15",
         "@web3-onboard/react": "^2.2.5",
+        "@web3-onboard/torus": "^2.2.0",
         "@web3-onboard/walletconnect": "^2.0.8",
         "@web3-onboard/walletlink": "^2.0.6",
         "amplitude-js": "^8.21.2",
